django/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
black==21.7b0
braintree==4.13.1
cachetools==5.2.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.8
click==8.1.3
colorama==0.4.5
configparser==5.0.2
croniter==1.3.4
cypress-django==0.1.0
Cython==0.29.24
cytoolz==0.12.0
dataclasses==0.6
ddtrace==0.56.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.8
django-bootstrap4==22.1
django-ckeditor==6.5.0
django-cors-headers==3.10.0
django-crum==0.7.9
django-debug-toolbar==3.4.0
django-embed-video==1.4.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==22.1
django-htmx==1.9.0
django-js-asset==2.0.0
django-redis==5.0.0
django-rq==2.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ERPpeek==1.7.1
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fakeredis==1.8
filelock==3.6.0
frozenlist==1.3.1
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
hexbytes==0.3.0
identify==2.5.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.15.0
keyboard==0.13.5
lru-dict==1.1.8
Markdown==3.4.1
MouseInfo==0.1.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.1
netaddr==0.8.0
nodeenv==1.6.0
numpy==1.21.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
parsimonious==0.8.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
pre-commit==2.18.1
protobuf==3.20.1
psycopg2-binary==2.9.3
py-spy==0.3.11
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyclean==2.2.0
pycryptodome==3.15.0
pyfiglet==0.8.post1
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
pysodium==0.7.11
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.2
pytweening==1.0.4
pytz==2021.3
pywin32==304
PyYAML==6.0
redis==4.2.2
regex==2022.7.25
requests==2.26.0
requests-oauthlib==1.3.1
rlp==2.0.1
rq==1.10.1
rq-scheduler==0.11.0
rsa==4.7.2
selenium==3.141.0
sentry-sdk==1.5.0
shell==1.0.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Telethon==1.23.0
Telethon-aio==1.0
tenacity==8.0.1
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
tzdata==2022.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.14.1
web3==5.30.0
webdriver-manager==3.8.2
websockets==9.1
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.8.1
